Add files via upload
The report is aimed at investigating through historical dataset indicating employee’s satisfaction and other conditions associated with each employee as at the last time the test was taken. This dataset is to help provide answers to what has been the factor(s) influencing employee satisfaction level on the job.